+++ resources/views/components/favicons.blade.php
@@ -1,4 +1,4 @@
-<link rel="manifest" href='data:application/manifest+json,{"name":"Nettoweb CMS","icons":[{"src":"{{ NETTO_CDN_URL.'/images/icon-192.png' }}","type":"image/png","sizes":"192x192"},{"src":"{{ NETTO_CDN_URL.'/images/icon-512.png' }}","type":"image/png","sizes":"512x512"}]}' />
-<link rel="icon" href="{{ NETTO_CDN_URL.'/images/favicon.ico' }}" sizes="any">
-<link rel="icon" href="{{ NETTO_CDN_URL.'/images/icon.svg' }}" type="image/svg+xml">
-<link rel="apple-touch-icon" href="{{ NETTO_CDN_URL.'/images/apple-touch-icon.png' }}">
+<link rel="manifest" href='data:application/manifest+json,{"name":"Nettoweb CMS","icons":[{"src":"{{ config('cms.assets.host').'/images/icon-192.png' }}","type":"image/png","sizes":"192x192"},{"src":"{{ config('cms.assets.host').'/images/icon-512.png' }}","type":"image/png","sizes":"512x512"}]}' />
+<link rel="icon" href="{{ config('cms.assets.host').'/images/favicon.ico' }}" sizes="any">
+<link rel="icon" href="{{ config('cms.assets.host').'/images/icon.svg' }}" type="image/svg+xml">
+<link rel="apple-touch-icon" href="{{ config('cms.assets.host').'/images/apple-touch-icon.png' }}">